refactor: move configs to src dir
