+++ .husky/pre-commit
@@ -1,1 +1,2 @@
+export NODE_OPTIONS="--experimental-strip-types"
 npx nano-staged
