--- index.js
@@ -1,4 +1,0 @@
-import baseConfig from './base.js';
-import typescriptConfig from './typescript.js';
-
-export default [...baseConfig, ...typescriptConfig];
