+++ src/configs/prettier.js
