+++ src/configs/server.js
@@ -1,7 +1,7 @@
 import securityPlugin from 'eslint-plugin-security';
 import globals from 'globals';
 
-import {ERROR, WARNING} from './constants.js';
+import {ERROR, WARNING} from '../constants.js';
 
 export default [
     {
