+++ src/index.js
@@ -1,0 +1,4 @@
+import baseConfig from './configs/base.js';
+import typescriptConfig from './configs/typescript.js';
+
+export default [...baseConfig, ...typescriptConfig];
