+++ test/client/eslint.config.js
@@ -1,8 +1,8 @@
-import a11yConfig from '../../a11y.js';
-import client from '../../client.js';
 import importConfig from '../../import-order.js';
-import mainConfig from '../../index.js';
-import prettierConfig from '../../prettier.js';
+import a11yConfig from '../../src/configs/a11y.js';
+import client from '../../src/configs/client.js';
+import prettierConfig from '../../src/configs/prettier.js';
+import mainConfig from '../../src/index.js';
 
 export default [
     ...mainConfig,
